mmap: extract into module/mmap/ following the _socket pattern
Moves the mmap class + module-level init from importing.rs into
module/mmap/{mod,moduledef,interp_mmap}.rs, mirroring the
pypy/module/mmap/ layout.  The previous init_mmap is renamed to
register_module and invoked from moduledef::init; init_mmap_type and
helpers stay private to interp_mmap.

importing.rs shrinks by 762 lines.
